refactor: Archive 페이지 초기 페이지 진입 속도 개선
